+++ src/test/scala/org/podval/xml/XmlNamespaceSpec.scala
@@ -129,7 +129,7 @@ final class XmlNamespaceSpec extends AnyFunSuite:
     assert(xml.name.localName == "p")
     assert(xml.name.prefix.contains("tei"))
     assert(xml.name.namespace.contains(tei))
-    assert(XmlExpandedName.asPairs(xml.getExpandedAttributes) == Seq("xmlns:tei" -> tei, "xml:id" -> "n1"))
+    assert(XmlExpandedName.asPairs(xml.getAttributes) == Seq("xmlns:tei" -> tei, "xml:id" -> "n1"))
     assert(attrName(xml, "xml:id").namespace.contains(XmlNamespace.xml.uri))
     assert(attrName(xml, "xmlns:tei").namespace.contains(XmlNamespace.xmlns.uri))
   }
@@ -141,7 +141,7 @@ final class XmlNamespaceSpec extends AnyFunSuite:
     assert(attrName(xml, "xmlns").namespace.contains(XmlNamespace.xmlns.uri))
   }
 
-  test("Xml rename and setAttributes keep an inherited namespace") {
+  test("Xml rename and set keep an inherited namespace") {
     val child: Xml.Element = children(parse(s"""<outer xmlns="$tei"><inner n="1"/></outer>""")).head
     assert(child.name.namespace.contains(tei))
     val renamed: Xml.Element = child.rename("p")
@@ -164,7 +164,7 @@ final class XmlNamespaceSpec extends AnyFunSuite:
     assert(el.label == "p")
     assert(el.scope.getURI("tei") == tei)
     assert(el.scope.getURI("xml") == XmlNamespace.xml.uri)
-    assert(XmlExpandedName.asPairs(ScalaXml.getExpandedAttributes(el)) == Seq("xmlns:tei" -> tei, "xml:id" -> "n1"))
+    assert(XmlExpandedName.asPairs(ScalaXml.getAttributes(el)) == Seq("xmlns:tei" -> tei, "xml:id" -> "n1"))
   }
 
   test("ScalaXml default xmlns is on scope") {